backend/requirements.txt
Flask
flask-marshmallow
Flask-Sqlalchemy
marshmallow-sqlalchemy
psycopg2-binary
python-dotenv
flask-cors